Removed LocalUserViewModel, called API directly from DonationDialog
